tasks and utils from models to the correct places
at least some of them
